+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Volume widget:
 - Add a "List connectors" link to each volume, to show all connectors in a
   volume bounding box.
 
+- The new "Min skeleton nodes" and "Min skeleton length" options allow to
+  further constrain volume based skeleton selections ("List skeletons").
+
 Miscellaneous:
 
 - The behavior of the Ctrl modifier on section navigation with , and . can now